sdfsdsfa

+++ src/main/java/team/gutterteam123/ledanimation/elemets/Controllable.java
@@ -2,7 +2,7 @@ package team.gutterteam123.ledanimation.elemets;
 
 public interface Controllable {
 
-    void setChannel(ChannelType channel, float value);
+    void setChannel(ChannelType channel, short value);
 
     String displayName();
